deep-ai/requirements.txt
fastapi==0.110.1
langchain==0.1.14
langchain-community==0.0.31
langchain-experimental==0.0.56
langchainhub==0.1.15
langchain-google-genai==1.0.3
langchain_openai==0.1.1
langchain-anthropic==0.1.4
python-dotenv==1.0.1
openai==1.16.0
tiktoken==0.6.0
python-multipart==0.0.9
unstructured==0.13.0
pytesseract==0.3.10
uvicorn==0.29.0
wikipedia==1.4.0
pydantic==2.6.4
pydantic_settings==2.2.1
faiss-cpu==1.8.0
pymysql
tabulate==0.9.0
starlette==0.37.2
python-jose==3.3.0
passlib==1.7.4
jose==1.0.0
bcrypt==4.1.2
colorlog==6.8.2
requests
redis==5.0.3
sqlmodel==0.0.16
SQLAlchemy==2.0.29
numpy==1.26.4
tokenizers==0.15.2
yagmail~=0.15.293
keyring==25.0.0
keyrings.alt==5.0.1
pydub~=0.25.1
edge-tts==6.1.10
python-docx
cryptography==42.0.5
google-generativeai==0.5.2
pypdf==4.1.0
dashscope==1.17.0
httpx~=0.26.0
